[core] feat(Utils): elementIsTextInput utility function
